combine osquery
